+++ tests/test_export_airflow.py
@@ -32,7 +32,7 @@ def test_generate_valid_airflow_dags(name, tmp_projects):
     subprocess.run(['python', f'dags/{name}.py'], check=True)
 
 
-def test_export_airflow(monkeypatch, tmp_sample_project):
+def test_export_airflow_sample_project(monkeypatch, tmp_sample_project):
     export.project(project_root='.', output_path='exported')
     monkeypatch.syspath_prepend('exported/dags')
     airflow_home = Path(tmp_sample_project, 'exported').resolve()
@@ -49,6 +49,23 @@ def test_export_airflow(monkeypatch, tmp_sample_project):
                 'clean': {'raw'},
                 'plot': {'clean'}
             }
+
+
+def test_export_airflow_custom_args(monkeypatch, tmp_sample_project):
+    Path('soopervisor.yaml').write_text('args: --some-arg')
+
+    export.project(project_root='.', output_path='exported')
+    monkeypatch.syspath_prepend('exported/dags')
+    airflow_home = Path(tmp_sample_project, 'exported').resolve()
+    monkeypatch.setenv('AIRFLOW_HOME', airflow_home)
+    mod = importlib.import_module('sample_project')
+    dag = mod.dag
+
+    assert 'ploomber task raw --some-arg' in dag.task_dict['raw'].bash_command
+    assert 'ploomber task clean --some-arg' in dag.task_dict[
+        'clean'].bash_command
+    assert 'ploomber task plot --some-arg' in dag.task_dict[
+        'plot'].bash_command
 
 
 def test_export_airflow_callables(monkeypatch, tmp_callables):